+++ js/util.js
@@ -1,0 +1,3 @@
+const getRandomInteger = (min, max) => Math.floor(min + Math.random() * (max + 1 - min));
+
+export {getRandomInteger};
